Autoclippify cargo-pgrx and other build-support crates (#1382)
Last round of autoclippifying things for now...

...mostly because pgrx-sql-entity-graph induces such massive errors in
`cargo clippy --fix` that it repeatedly aborts and emits the most
fascinating errors, rather than try to complete the automatic fixes.

+++ pgrx-examples/errors/src/lib.rs
@@ -14,7 +14,7 @@ pgrx::pg_module_magic!();
 
 #[pg_extern]
 fn array_with_null_and_panic(input: Vec<Option<i32>>) -> i64 {
-    let mut sum = 0 as i64;
+    let mut sum = 0_i64;
 
     for i in input {
         sum += i.expect("NULL elements in input array are not supported") as i64;
